+++ src/main/java/com/livinglogic/ul4/BoundArguments.java
@@ -14,7 +14,7 @@ import java.util.Iterator;
 import static java.util.Arrays.asList;
 
 
-public class BoundArguments implements AutoCloseable
+public class BoundArguments
 {
 	protected String name;
 	protected Signature signature;
@@ -497,41 +497,41 @@ public class BoundArguments implements AutoCloseable
 		return makeIterator(argumentsByName.get(name), -1, name);
 	}
 
-	/**
-	"Destroys" a {@code BoundArguments} object to simplify the work the Java GC has to do
-	After the call the object is no longer usable.
+	// /**
+	// "Destroys" a {@code BoundArguments} object to simplify the work the Java GC has to do
+	// After the call the object is no longer usable.
 
-	Note that this should not be called for any template arguments,
-	as the variables might be referenced by a closure.
-	However for {@code Function} objects that can be done without problems.
-	**/
-	@Override
-	public void close()
-	{
-		if (signature != null)
-		{
-			if (signature.hasVarKeyword())
-			{
-				Map<String, Object> kwargs = (Map<String, Object>)get(argumentsByPosition.length - 1);
-				kwargs.clear();
-			}
-			if (signature.hasVarPositional())
-			{
-				List<Object> args = (List<Object>)get(argumentsByPosition.length - (signature.hasVarKeyword() ? 2 : 1));
-				args.clear();
-			}
-			argumentsByPosition = null;
-			if (argumentsByName != null)
-			{
-				argumentsByName.clear();
-				argumentsByName = null;
-			}
-			signature = null;
-		}
-		else
-		{
-			argumentsByName.clear();
-			argumentsByName = null;
-		}
-	}
+	// Note that this should not be called for any template arguments,
+	// as the variables might be referenced by a closure.
+	// However for {@code Function} objects that can be done without problems.
+	// **/
+	// @Override
+	// public void close()
+	// {
+	// 	if (signature != null)
+	// 	{
+	// 		if (signature.hasVarKeyword())
+	// 		{
+	// 			Map<String, Object> kwargs = (Map<String, Object>)get(argumentsByPosition.length - 1);
+	// 			kwargs.clear();
+	// 		}
+	// 		if (signature.hasVarPositional())
+	// 		{
+	// 			List<Object> args = (List<Object>)get(argumentsByPosition.length - (signature.hasVarKeyword() ? 2 : 1));
+	// 			args.clear();
+	// 		}
+	// 		argumentsByPosition = null;
+	// 		if (argumentsByName != null)
+	// 		{
+	// 			argumentsByName.clear();
+	// 			argumentsByName = null;
+	// 		}
+	// 		signature = null;
+	// 	}
+	// 	else
+	// 	{
+	// 		argumentsByName.clear();
+	// 		argumentsByName = null;
+	// 	}
+	// }
 }