+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Performance optimizations:
   non-existant attributes indistinguishable from attributes that happen to
   be ``UndefinedAttribute`` objects without calling ``hasAttrUL4()``, but this
   distinction can't be made in UL4 anyway (without calling ``hasattr()``).
+* ``BoundArguments`` objects are now no longer autocloseable. Instead we let
+  the garbage collector clean them up naturally.
 
 
 exp-170 (2022-03-30)